requirements.txt
black==24.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
filelock==3.13.1
fonttools==4.49.0
fsspec==2024.2.0
idna==3.6
isort==5.13.2
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
mpmath==1.3.0
mypy==1.8.0
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.4
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
scipy==1.12.0
six==1.16.0
sympy==1.12
torch==2.2.1
torchvision==0.17.1
tornado==6.4
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
visdom==0.2.4
websocket-client==1.7.0
